pyproject.toml
[tool.poetry]
name = "dynamic-neural-network-refinement"
version = "0.2.1"
description = "A project for dynamic neural network refinement with state-of-the-art practices."
authors = ["Reece Dixon <qtt@null.net>"]
license = "GNU Affero General Public License v3.0"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9 || ^3.10 || ^3.11"
torch = ">=2.0.0"
torchvision = ">=0.15.1"
PyYAML = ">=6.0"
wandb = "^0.12.21"
loguru = "^0.5.3"
matplotlib = ">=3.5.3"
seaborn = ">=0.12.2"
pandas = "^1.3.5"
plotly = "^5.18.0"
ray = { version = "^2.0.0", extras = ["tune"] }
flake8 = "^6.1.0"
black = "^23.7.0"
pytest = "^7.4.2"
great_expectations = "^0.16.0"
dvc = "^2.53.0"
fastapi = ">=0.99.0"
uvicorn = ">=0.23.0"
json-log-formatter = "^0.5.0"
numpy = ">=1.21.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
